// src/core/contract/domain/models/filter_contract_model.ts
import type { Id } from "@/src/common/utils/types";
import type { ContractStatus } from "@/src/core/app/enums/contract_status";

export class FilterContractModel {
  id?: Id;
  isActive?: boolean;
  status?: ContractStatus;
  preferredStartDate?: string;
  startDate?: string;
  expectedEndDate?: string;
  endDate?: string;
  createdAt?: string;
  power1?: number;
  power2?: number;
  power3?: number;
  power4?: number;
  power5?: number;
  power6?: number;
  period?: number;
  signatureFirstName?: string;
  signatureLastName?: string;
  signatureDni?: string;
  signatureEmail?: string;
  signaturePhone?: string;
  client?: string;
  cif?: string;
  cups?: string;
  supplyAlias?: string;
  supplyAddress?: string;
  supplyCity?: string;
  supplyProvince?: string;
  rateType?: string;
  rateName?: string;
  user?: string;

  constructor(contract: {
    id?: Id;
    isActive?: boolean;
    status?: ContractStatus;
    preferredStartDate?: string;
    startDate?: string;
    expectedEndDate?: string;
    endDate?: string;
    createdAt?: string;
    power1?: number;
    power2?: number;
    power3?: number;
    power4?: number;
    power5?: number;
    power6?: number;
    period?: number;
    signatureFirstName?: string;
    signatureLastName?: string;
    signatureDni?: string;
    signatureEmail?: string;
    signaturePhone?: string;
    client?: string;
    cif?: string;
    cups?: string;
    supplyAlias?: string;
    supplyAddress?: string;
    supplyCity?: string;
    supplyProvince?: string;
    rateType?: string;
    rateName?: string;
    user?: string;
  }) {
    this.id = contract.id;
    this.isActive = contract.isActive;
    this.status = contract.status;
    this.preferredStartDate = contract.preferredStartDate;
    this.startDate = contract.startDate;
    this.expectedEndDate = contract.expectedEndDate;
    this.endDate = contract.endDate;
    this.createdAt = contract.createdAt;
    this.power1 = contract.power1;
    this.power2 = contract.power2;
    this.power3 = contract.power3;
    this.power4 = contract.power4;
    this.power5 = contract.power5;
    this.power6 = contract.power6;
    this.period = contract.period;
    this.signatureFirstName = contract.signatureFirstName;
    this.signatureLastName = contract.signatureLastName;
    this.signatureDni = contract.signatureDni;
    this.signatureEmail = contract.signatureEmail;
    this.signaturePhone = contract.signaturePhone;
    this.client = contract.client;
    this.cif = contract.cif;
    this.cups = contract.cups;
    this.supplyAlias = contract.supplyAlias;
    this.supplyAddress = contract.supplyAddress;
    this.supplyCity = contract.supplyCity;
    this.supplyProvince = contract.supplyProvince;
    this.rateType = contract.rateType;
    this.rateName = contract.rateName;
    this.user = contract.user;
  }
}